Make it easier to run and debug the test suite
We _ought_ to add node-inspector to the devDependencies to enable the test:debug script, but that would currently break `npm install` for folks on node v7 because of  https://github.com/node-inspector/node-inspector/issues/950.  For now, contributors will have to install node v6 or lower and manually `npm install node-inspector`.